+++ tests/pda-derivation/tests/typescript.spec.ts
@@ -128,9 +128,20 @@ describe("typescript", () => {
       // intentionally not provided to test the error message
       await program.methods.resolutionError().rpc();
       throw new Error("Should throw due to account resolution failure!");
-    } catch (e) {
+    } catch (e: any) {
       expect(e.message).to.equal(
         "Reached maximum depth for account resolution. Unresolved accounts: `pda`, `anotherPda`"
+      );
+    }
+  });
+
+  it("Throws a dedicated error for self-referencing PDAs", async () => {
+    try {
+      await program.methods.selfReferencingSeed().rpc();
+      throw new Error("Should throw due to account resolution failure!");
+    } catch (e: any) {
+      expect(e.message).to.equal(
+        "Self-referencing PDAs must be provided manually. Unresolved PDAs: `selfReferencingPda`"
       );
     }
   });